fix | sprint2 | FRB-188 | MQTT 연결 관련해서 재연결 로직 추가 | 윤다인

+++ src/main/java/com/factoreal/backend/domain/controlLog/entity/ControlLog.java
@@ -25,7 +25,7 @@ public class ControlLog {
     private String controlType;
 
     @Column(name = "control_val")
-    private Integer controlVal;
+    private Double controlVal;
 
     @Column(name = "control_stat")
     private Integer controlStat;
@@ -43,6 +43,5 @@ public class ControlLog {
     @ManyToOne(fetch = FetchType.LAZY)
     @JoinColumn(name = "zone_id", referencedColumnName = "zone_id")
     private Zone zone;
-
 
 }
